tests: consolidate matrix test infrastructure
Consolidate matrix test infrastructure so that list objects matrix
test can also use some of the check matrix infra.
