+++ tests/tests.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"testing"
+	"time"
 
 	"google.golang.org/grpc"
 
@@ -23,6 +24,15 @@ type TestClientBootstrapper interface {
 	CreateStore(ctx context.Context, in *openfgav1.CreateStoreRequest, opts ...grpc.CallOption) (*openfgav1.CreateStoreResponse, error)
 	WriteAuthorizationModel(ctx context.Context, in *openfgav1.WriteAuthorizationModelRequest, opts ...grpc.CallOption) (*openfgav1.WriteAuthorizationModelResponse, error)
 	Write(ctx context.Context, in *openfgav1.WriteRequest, opts ...grpc.CallOption) (*openfgav1.WriteResponse, error)
+}
+
+// ClientInterface defines client interface for running tests.
+type ClientInterface interface {
+	TestClientBootstrapper
+	Check(ctx context.Context, in *openfgav1.CheckRequest, opts ...grpc.CallOption) (*openfgav1.CheckResponse, error)
+	ListUsers(ctx context.Context, in *openfgav1.ListUsersRequest, opts ...grpc.CallOption) (*openfgav1.ListUsersResponse, error)
+	ListObjects(ctx context.Context, in *openfgav1.ListObjectsRequest, opts ...grpc.CallOption) (*openfgav1.ListObjectsResponse, error)
+	StreamedListObjects(ctx context.Context, in *openfgav1.StreamedListObjectsRequest, opts ...grpc.CallOption) (openfgav1.OpenFGAService_StreamedListObjectsClient, error)
 }
 
 // StartServer calls StartServerWithContext. See the docs for that.
@@ -61,4 +71,26 @@ func StartServerWithContext(t testing.TB, cfg *serverconfig.Config, serverCtx *r
 	})
 
 	testutils.EnsureServiceHealthy(t, cfg.GRPC.Addr, cfg.HTTP.Addr, nil)
+}
+
+// BuildClientInterface sets up test client interface to be used for matrix test.
+func BuildClientInterface(t *testing.T, engine string, experimentals []string) ClientInterface {
+	cfg := serverconfig.MustDefaultConfig()
+	if len(experimentals) > 0 {
+		cfg.Experimentals = append(cfg.Experimentals, experimentals...)
+	}
+	cfg.Log.Level = "error"
+	cfg.Datastore.Engine = engine
+	cfg.ListUsersDeadline = 0   // no deadline
+	cfg.ListObjectsDeadline = 0 // no deadline
+	// extend the timeout for the tests, coverage makes them slower
+	cfg.RequestTimeout = 10 * time.Second
+
+	cfg.CheckIteratorCache.Enabled = true
+	cfg.ContextPropagationToDatastore = true
+
+	StartServer(t, cfg)
+
+	conn := testutils.CreateGrpcConnection(t, cfg.GRPC.Addr)
+	return openfgav1.NewOpenFGAServiceClient(conn)
 }
